+++ .cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+target-dir = ".build"
