--- .hypergit
@@ -1,3 +1,0 @@
-@GENERATED 1680892544
-USER=hypernite@Hypernite
-HOST=git.linkrbot.com
