+++ rektor/cli/src/cli.rs
@@ -1,0 +1,7 @@
+pub struct CLI;
+
+impl CLI {
+    pub fn run() {
+        unimplemented!("Rektor DB cli not implemented");
+    }
+}
