+++ rektor/cli/src/main.rs
@@ -1,0 +1,7 @@
+use cli::CLI;
+
+mod cli;
+
+fn main() {
+    CLI::run();
+}
