+++ rektor/core/src/lib.rs
@@ -1,0 +1,19 @@
+//! # Core
+
+#![deny(clippy::all)]
+
+/// Adds two numbers together.
+pub fn add(left: usize, right: usize) -> usize {
+    left + right
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn it_works() {
+        let result = add(2, 2);
+        assert_eq!(result, 4);
+    }
+}
